Added JWT_SECRET into a middleware file so as to be accessible into other js files
